Move over tests, rename zarr.rs to store.rs
