+++ icechunk/src/session.rs
@@ -19,20 +19,15 @@ use crate::{
         manifest::{
             ChunkInfo, ChunkPayload, ChunkRef, Manifest, ManifestExtents, ManifestRef,
             VirtualChunkRef, VirtualReferenceError,
-        },
-        snapshot::{
+        }, snapshot::{
             ManifestFileInfo, NodeData, NodeSnapshot, NodeType, Snapshot,
             SnapshotProperties, UserAttributesSnapshot, ZarrArrayMetadata,
-        },
-        transaction_log::TransactionLog,
-        ByteRange, ChunkIndices, IcechunkFormatError, ManifestId, NodeId, Path,
-        SnapshotId,
+        }, transaction_log::TransactionLog, ByteRange, ChunkIndices, IcechunkFormatError, ManifestId, NodeId, ObjectId, Path, SnapshotId
     },
     metadata::UserAttributes,
     refs::{fetch_branch_tip, update_branch, RefError},
-    repo::{RepositoryConfig, RepositoryError},
+    repository::{RepositoryConfig, RepositoryError},
     storage::virtual_ref::{construct_valid_byte_range, VirtualChunkResolver},
-    zarr::ObjectId,
     Storage, StorageError,
 };
